+++ router.c
@@ -425,6 +425,7 @@ router_readconfig(router *orig,
 		const char *path,
 		size_t queuesize,
 		size_t batchsize,
+		int maxstalls,
 		unsigned short iotimeout)
 {
 	FILE *cnf;
@@ -811,7 +812,7 @@ router_readconfig(router *orig,
 								*proto == 'f' ? CON_FILE :
 								*proto == 'u' ? CON_UDP : CON_TCP,
 								walk == (void *)1 ? NULL : walk,
-								queuesize, batchsize, iotimeout);
+								queuesize, batchsize, maxstalls, iotimeout);
 						if (newserver == NULL) {
 							logerr("failed to add server %s:%d (%s) "
 									"to cluster %s: %s\n", ip, port, proto,
@@ -1897,7 +1898,7 @@ router_readconfig(router *orig,
 				for (i = 0; i < globbuf.gl_matchc; i++) {
 					globpath = globbuf.gl_pathv[i];
 					ret = router_readconfig(ret, globpath, queuesize,
-							batchsize, iotimeout);
+							batchsize, maxstalls, iotimeout);
 					if (ret == NULL)
 						break;
 				}
@@ -1908,7 +1909,7 @@ router_readconfig(router *orig,
 			} else {
 				/* include path is a regular file path */
 				ret = router_readconfig(ret, name, queuesize,
-						batchsize, iotimeout);
+						batchsize, maxstalls, iotimeout);
 			}
 			if (ret == NULL)
 				/* router_readconfig already barked and freed ret */